I used the Mailkit method in the ForgetPassord section
