+++ Core/Service/Authservice.cs
@@ -15,7 +15,7 @@ using System.Text;
 
 namespace Service
 {
-    public class Authservice(UserManager<AppUsers>userManager,IOptions<JWTOptions>options,IMapper mapper, SignInManager<AppUsers> signInManager, IOptions<MailSetting>options1) : IAuthservice
+    public class Authservice(UserManager<AppUsers>userManager,IOptions<JWTOptions>options,IMapper mapper, SignInManager<AppUsers> signInManager) : IAuthservice
     {
         public async Task<bool> CheckEmailExistAsync(string email)
         {
@@ -192,19 +192,5 @@ namespace Service
             return "Password has been reset successfully.";
         }
 
-        public async Task SendEmailAsync(Email email)
-        {
-            var mail=new MimeMessage();
-            mail.Subject = email.Subject;
-            mail.From.Add(new MailboxAddress(options1.Value.DisplayName, options1.Value.Email));
-            mail.To.Add(MailboxAddress.Parse(email.To));
-            var builder = new BodyBuilder();
-            builder.HtmlBody = email.Body;
-            mail.Body = builder.ToMessageBody();
-            using var smtp=new MailKit.Net.Smtp.SmtpClient();
-            smtp.Connect(options1.Value.Host, options1.Value.Port, MailKit.Security.SecureSocketOptions.StartTls);
-            smtp.Authenticate(options1.Value.Email, options1.Value.Password);
-            smtp.Send(mail);
-        }
     }
 }
