+++ Core/ServiceAbstraction/IMailServices.cs
@@ -1,0 +1,14 @@
+﻿using Domain.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ServiceAbstraction
+{
+    public interface IMailServices
+    {
+        Task SendEmailAsync(Email email);
+    }
+}
